Move architecture-specific code to src/
