+++ amalgamation.sh
@@ -17,11 +17,23 @@ $SCRIPTPATH/src/simdjson.cpp
 $SCRIPTPATH/src/jsonioutil.cpp
 $SCRIPTPATH/src/jsonminifier.cpp
 $SCRIPTPATH/src/jsonparser.cpp
+$SCRIPTPATH/src/simd_input.h
+$SCRIPTPATH/src/simd_input_haswell.h
+$SCRIPTPATH/src/simd_input_westmere.h
+$SCRIPTPATH/src/simd_input_arm64.h
+$SCRIPTPATH/src/simdutf8check.h
+$SCRIPTPATH/src/simdutf8check_haswell.h
+$SCRIPTPATH/src/simdutf8check_westmere.h
+$SCRIPTPATH/src/simdutf8check_arm64.h
 $SCRIPTPATH/src/stage1_find_marks_westmere.h
 $SCRIPTPATH/src/stage1_find_marks_haswell.h
 $SCRIPTPATH/src/stage1_find_marks_arm64.h
 $SCRIPTPATH/src/stage1_find_marks.cpp
 $SCRIPTPATH/src/stage2_build_tape.cpp
+$SCRIPTPATH/src/stringparsing.h
+$SCRIPTPATH/src/stringparsing_westmere.h
+$SCRIPTPATH/src/stringparsing_haswell.h
+$SCRIPTPATH/src/stringparsing_arm64.h
 $SCRIPTPATH/src/parsedjson.cpp
 $SCRIPTPATH/src/parsedjsoniterator.cpp
 "
@@ -38,22 +50,10 @@ $SCRIPTPATH/include/simdjson/jsoncharutils.h
 $SCRIPTPATH/include/simdjson/jsonformatutils.h
 $SCRIPTPATH/include/simdjson/jsonioutil.h
 $SCRIPTPATH/include/simdjson/simdprune_tables.h
-$SCRIPTPATH/include/simdjson/simd_input.h
-$SCRIPTPATH/include/simdjson/simd_input_haswell.h
-$SCRIPTPATH/include/simdjson/simd_input_westmere.h
-$SCRIPTPATH/include/simdjson/simd_input_arm64.h
-$SCRIPTPATH/include/simdjson/simdutf8check.h
-$SCRIPTPATH/include/simdjson/simdutf8check_haswell.h
-$SCRIPTPATH/include/simdjson/simdutf8check_westmere.h
-$SCRIPTPATH/include/simdjson/simdutf8check_arm64.h
 $SCRIPTPATH/include/simdjson/jsonminifier.h
 $SCRIPTPATH/include/simdjson/parsedjson.h
 $SCRIPTPATH/include/simdjson/parsedjsoniterator.h
 $SCRIPTPATH/include/simdjson/stage1_find_marks.h
-$SCRIPTPATH/include/simdjson/stringparsing.h
-$SCRIPTPATH/include/simdjson/stringparsing_westmere.h
-$SCRIPTPATH/include/simdjson/stringparsing_haswell.h
-$SCRIPTPATH/include/simdjson/stringparsing_arm64.h
 $SCRIPTPATH/include/simdjson/numberparsing.h
 $SCRIPTPATH/include/simdjson/stage2_build_tape.h
 $SCRIPTPATH/include/simdjson/jsonparser.h
