+++ src/simd_input.h
