+++ src/simdutf8check_westmere.h
@@ -2,7 +2,7 @@
 #define SIMDJSON_SIMDUTF8CHECK_WESTMERE_H
 
 #include "simdjson/portability.h"
-#include "simdjson/simdutf8check.h"
+#include "simdutf8check.h"
 #include <stddef.h>
 #include <stdint.h>
 #include <string.h>
