+++ src/stage2_build_tape_common.h
